docs: Partially updated documentation. \n fix: Replaced all raw pointers with smart ones. Fixed bugs in the BinaryStream class. \n feat: The server now handles the standard RakNet handshake. Synchronized threads with std::mutex. Now the log has become more beautiful! Added thread manager.

+++ src/network/protocol/EncapsulatedPacket.h
@@ -26,7 +26,7 @@ public:
 	void decode(PacketSerializer* out) {
 		uint8_t flags = out->readByte();
 		this->reliability = (flags & (0b111 << 5)) >> 5;
-		this->hasSplit = (flags & 0b00010000) == 1 ? true : false;
+		this->hasSplit = (int) (flags & 0b00010000) == 1 ? true : false;
 		
 		this->length = out->readUShort() >> 3;
